+++ src/main/java/gift/entity/Product/ProductName.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.entity.Product;
 
 import gift.common.exception.core.CustomException;
 import jakarta.persistence.Column;
