+++ src/main/java/gift/entity/Product/ProductPrice.java
@@ -1,4 +1,4 @@
-package gift.entity;
+package gift.entity.Product;
 
 import gift.common.exception.core.CustomException;
 import jakarta.persistence.Column;
