--- src/main/java/gift/entity/Wish.java
@@ -1,61 +1,0 @@
-package gift.entity;
-
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "wish")
-public class Wish {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    @Embedded
-    private WishQuantity quantity;
-
-    protected Wish() {
-    }
-
-    private Wish(Long id, Member member, Product product, Integer quantity) {
-        this.id = id;
-        this.member = member;
-        this.product = product;
-        this.quantity = new WishQuantity(quantity);
-    }
-
-    public Wish(Member member, Product product, Integer quantity) {
-        this(null, member, product, quantity);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public int getQuantity() {
-        return quantity.quantity();
-    }
-}
