+++ src/main/java/gift/external/KaKaoMessageClient.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.common.code.CustomResponseCode;
 import gift.common.exception.KaKaoClientException;
 import gift.common.exception.UnauthorizedException;
-import gift.entity.Member;
-import gift.entity.Order;
-import gift.entity.Product;
-import gift.entity.ProductOption;
+import gift.entity.Member.Member;
+import gift.entity.Order.Order;
+import gift.entity.Product.Option.ProductOption;
+import gift.entity.Product.Product;
 import gift.service.Auth.AuthService;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Value;
