+++ src/main/java/gift/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.entity.Order;
+import gift.entity.Order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
