+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.entity.Product;
+import gift.entity.Product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
