+++ src/main/java/gift/service/Order/OrderService.java
@@ -2,7 +2,7 @@ package gift.service.Order;
 
 import gift.dto.Order.OrderRequest;
 import gift.dto.Order.OrderResponse;
-import gift.entity.Member;
+import gift.entity.Member.Member;
 
 public interface OrderService {
