+++ src/main/java/gift/service/Order/OrderServiceImpl.java
@@ -4,9 +4,9 @@ import gift.common.code.CustomResponseCode;
 import gift.common.exception.core.CustomException;
 import gift.dto.Order.OrderRequest;
 import gift.dto.Order.OrderResponse;
-import gift.entity.Member;
-import gift.entity.Order;
-import gift.entity.ProductOption;
+import gift.entity.Member.Member;
+import gift.entity.Order.Order;
+import gift.entity.Product.Option.ProductOption;
 import gift.external.KaKaoMessageClient;
 import gift.repository.OptionRepository;
 import gift.repository.OrderRepository;
