+++ src/main/java/gift/service/Product/Option/OptionServiceImpl.java
@@ -4,8 +4,8 @@ import gift.common.code.CustomResponseCode;
 import gift.common.exception.core.CustomException;
 import gift.dto.Product.Option.ProductOptionRequest;
 import gift.dto.Product.Option.ProductOptionResponse;
-import gift.entity.Product;
-import gift.entity.ProductOption;
+import gift.entity.Product.Option.ProductOption;
+import gift.entity.Product.Product;
 import gift.repository.OptionRepository;
 import gift.repository.ProductRepository;
 import java.util.List;
