+++ src/main/java/gift/service/Wish/WishServiceImpl.java
@@ -8,10 +8,10 @@ import gift.dto.Pagination.Pagination;
 import gift.dto.Wish.WishRequest;
 import gift.dto.Wish.WishResponse;
 import gift.dto.Wish.WishSortField;
-import gift.entity.Member;
-import gift.entity.Product;
-import gift.entity.ProductOption;
-import gift.entity.Wish;
+import gift.entity.Member.Member;
+import gift.entity.Product.Option.ProductOption;
+import gift.entity.Product.Product;
+import gift.entity.Wish.Wish;
 import gift.repository.ProductRepository;
 import gift.repository.WishRepository;
 import org.springframework.data.domain.Page;
