+++ src/test/java/gift/e2e/OrderE2ETest.java
@@ -6,10 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import gift.auth.jwt.JwtUtil;
 import gift.common.code.CustomResponseCode;
 import gift.dto.Order.OrderRequest;
-import gift.entity.Member;
-import gift.entity.Product;
-import gift.entity.ProductOption;
-import gift.entity.Wish;
+import gift.entity.Member.Member;
+import gift.entity.Product.Option.ProductOption;
+import gift.entity.Product.Product;
+import gift.entity.Wish.Wish;
 import gift.external.KaKaoMessageClient;
 import gift.repository.MemberRepository;
 import gift.repository.OptionRepository;
