+++ src/test/java/gift/repository/MemberRepositoryTest.java
@@ -2,7 +2,7 @@ package gift.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.entity.Member;
+import gift.entity.Member.Member;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
