+++ src/test/java/gift/repository/WishRepositoryTest.java
@@ -3,9 +3,9 @@ package gift.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import gift.entity.Member;
-import gift.entity.Product;
-import gift.entity.Wish;
+import gift.entity.Member.Member;
+import gift.entity.Product.Product;
+import gift.entity.Wish.Wish;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
